feat: backend-service-refactor-service-package-002 - Extract local models package
