+++ pkg/localmodels/assets/puller_test.go
@@ -1,4 +1,4 @@
-package modelassets
+package assets
 
 import (
 	"context"
