+++ pkg/service/model_catalog.go
@@ -4,9 +4,6 @@ import (
 	"context"
 	"encoding/json"
 	"fmt"
-	"runtime"
-	"sort"
-	"strconv"
 	"strings"
 	"time"
 
@@ -14,77 +11,27 @@ import (
 	"github.com/portpowered/infinite-you/pkg/apisurface"
 	factoryconfig "github.com/portpowered/infinite-you/pkg/config"
 	"github.com/portpowered/infinite-you/pkg/interfaces"
+	"github.com/portpowered/infinite-you/pkg/localmodels"
 	"github.com/portpowered/infinite-you/pkg/logging"
-	"github.com/portpowered/infinite-you/pkg/service/modelassets"
 	"github.com/portpowered/infinite-you/pkg/workcontent"
 	"github.com/portpowered/infinite-you/pkg/workers"
 	workerexecutor "github.com/portpowered/infinite-you/pkg/workers/executor"
 )
 
-func (fs *FactoryService) ListModels(_ context.Context) (factoryapi.ListModelsResponse, error) {
-	catalog, err := fs.currentModelCatalog()
-	if err != nil {
-		return factoryapi.ListModelsResponse{}, err
-	}
-	results := make([]factoryapi.ModelSummary, 0, len(catalog))
-	for _, entry := range catalog {
-		results = append(results, entry.summary)
-	}
-	sort.Slice(results, func(i, j int) bool {
-		return results[i].Name < results[j].Name
-	})
-	return factoryapi.ListModelsResponse{Results: results}, nil
+func (fs *FactoryService) ListModels(ctx context.Context) (factoryapi.ListModelsResponse, error) {
+	_ = ctx
+	return localmodels.ListModels(fs.currentRuntimeConfig())
 }
 
-func (fs *FactoryService) GetModel(_ context.Context, modelName string) (factoryapi.ModelDetail, error) {
-	catalog, err := fs.currentModelCatalog()
-	if err != nil {
-		return factoryapi.ModelDetail{}, err
-	}
-	key := canonicalModelName(modelName)
-	if key == "" {
-		return factoryapi.ModelDetail{}, fmt.Errorf("%w: empty model name", apisurface.ErrModelNotFound)
-	}
-	entry, ok := catalog[key]
-	if !ok {
-		return factoryapi.ModelDetail{}, fmt.Errorf("%w: %s", apisurface.ErrModelNotFound, modelName)
-	}
-	return entry.detail, nil
+func (fs *FactoryService) GetModel(ctx context.Context, modelName string) (factoryapi.ModelDetail, error) {
+	_ = ctx
+	return localmodels.GetModel(fs.currentRuntimeConfig(), modelName)
 }
 
-type modelAssetPuller interface {
-	PullModel(ctx context.Context, runtimeCfg *factoryconfig.LoadedFactoryConfig, modelName string) (apisurface.ModelPullResult, error)
-	EnsureModelAvailable(ctx context.Context, runtimeCfg *factoryconfig.LoadedFactoryConfig, worker *interfaces.WorkerConfig) error
-	ResolveModelCache(ctx context.Context, runtimeCfg *factoryconfig.LoadedFactoryConfig, worker *interfaces.WorkerConfig) (localModelCacheLayout, error)
-}
-
-type modelAssetPullerAdapter struct {
-	inner *modelassets.Puller
-}
+type modelAssetPuller = localmodels.AssetPuller
 
 func newModelAssetPuller(cacheDir string) modelAssetPuller {
-	return modelAssetPullerAdapter{inner: modelassets.NewPuller(cacheDir, runtime.GOOS, runtime.GOARCH)}
-}
-
-func (p modelAssetPullerAdapter) PullModel(ctx context.Context, runtimeCfg *factoryconfig.LoadedFactoryConfig, modelName string) (apisurface.ModelPullResult, error) {
-	return p.inner.PullModel(ctx, runtimeCfg, modelName)
-}
-
-func (p modelAssetPullerAdapter) EnsureModelAvailable(ctx context.Context, runtimeCfg *factoryconfig.LoadedFactoryConfig, worker *interfaces.WorkerConfig) error {
-	return p.inner.EnsureModelAvailable(ctx, runtimeCfg, worker)
-}
-
-func (p modelAssetPullerAdapter) ResolveModelCache(ctx context.Context, runtimeCfg *factoryconfig.LoadedFactoryConfig, worker *interfaces.WorkerConfig) (localModelCacheLayout, error) {
-	layout, err := p.inner.ResolveModelCache(ctx, runtimeCfg, worker)
-	if err != nil {
-		return localModelCacheLayout{}, err
-	}
-	return localModelCacheLayout{
-		ModelName: layout.ModelName,
-		CachePath: layout.CachePath,
-		Revision:  layout.Revision,
-		Files:     layout.Files,
-	}, nil
+	return localmodels.NewAssetPuller(cacheDir)
 }
 
 func (fs *FactoryService) PullModel(ctx context.Context, modelName string) (apisurface.ModelPullResult, error) {
@@ -92,8 +39,8 @@ func (fs *FactoryService) PullModel(ctx context.Context, modelName string) (apis
 	if runtimeCfg == nil {
 		return apisurface.ModelPullResult{}, fmt.Errorf("factory service runtime is not available")
 	}
-	models := buildModelCatalog(runtimeCfg)
-	key := canonicalModelName(modelName)
+	models := localmodels.BuildCatalog(runtimeCfg)
+	key := localmodels.CanonicalModelName(modelName)
 	if key == "" {
 		return apisurface.ModelPullResult{}, fmt.Errorf("%w: empty model name", apisurface.ErrModelNotFound)
 	}
@@ -101,7 +48,7 @@ func (fs *FactoryService) PullModel(ctx context.Context, modelName string) (apis
 	if !ok {
 		return apisurface.ModelPullResult{}, fmt.Errorf("%w: %s", apisurface.ErrModelNotFound, modelName)
 	}
-	if entry.summary.ProviderLocality != factoryapi.WorkerModelLocalityLocal {
+	if entry.Summary.ProviderLocality != factoryapi.WorkerModelLocalityLocal {
 		return apisurface.ModelPullResult{}, fmt.Errorf("%w: model %q is not a local model", apisurface.ErrModelPullUnsupported, modelName)
 	}
 	return fs.modelAssetPuller().PullModel(ctx, runtimeCfg, modelName)
@@ -120,422 +67,6 @@ func (fs *FactoryService) modelAssetPuller() modelAssetPuller {
 		fs.modelAssets = puller
 	}
 	return puller
-}
-
-type discoveredModelCatalogEntry struct {
-	summary factoryapi.ModelSummary
-	detail  factoryapi.ModelDetail
-}
-
-type discoveredModelAggregate struct {
-	name           string
-	locality       string
-	localities     map[string]struct{}
-	workerCount    int
-	localCount     int
-	cloudCount     int
-	operations     map[string]factoryapi.ModelOperation
-	modalities     map[factoryapi.ModelOperationContentType]struct{}
-	resources      map[string]factoryapi.ModelResourceSummary
-	capabilities   []factoryapi.ModelCapability
-	workerNames    []string
-	hasAnyResource bool
-	hasModelScoped bool
-}
-
-func (fs *FactoryService) currentModelCatalog() (map[string]discoveredModelCatalogEntry, error) {
-	if fs == nil {
-		return nil, fmt.Errorf("factory service is required")
-	}
-	runtimeCfg := fs.currentRuntimeConfig()
-	if runtimeCfg == nil {
-		return nil, fmt.Errorf("factory service runtime is not available")
-	}
-	return buildModelCatalog(runtimeCfg), nil
-}
-
-func buildModelCatalog(runtimeCfg *factoryconfig.LoadedFactoryConfig) map[string]discoveredModelCatalogEntry {
-	if runtimeCfg == nil || runtimeCfg.FactoryConfig() == nil {
-		return map[string]discoveredModelCatalogEntry{}
-	}
-
-	factoryCfg := runtimeCfg.FactoryConfig()
-	resourceByName := make(map[string]interfaces.ResourceConfig, len(factoryCfg.Resources))
-	for _, resource := range factoryCfg.Resources {
-		resourceByName[resource.Name] = resource
-	}
-
-	aggregates := make(map[string]*discoveredModelAggregate)
-	for _, worker := range factoryCfg.Workers {
-		if strings.TrimSpace(worker.Type) != interfaces.WorkerTypeModel {
-			continue
-		}
-		key := canonicalModelName(worker.Model)
-		if key == "" {
-			continue
-		}
-		aggregate := aggregates[key]
-		if aggregate == nil {
-			aggregate = &discoveredModelAggregate{
-				name:       strings.TrimSpace(worker.Model),
-				locality:   strings.TrimSpace(worker.ModelLocality),
-				localities: make(map[string]struct{}),
-				operations: make(map[string]factoryapi.ModelOperation),
-				modalities: make(map[factoryapi.ModelOperationContentType]struct{}),
-				resources:  make(map[string]factoryapi.ModelResourceSummary),
-			}
-			aggregates[key] = aggregate
-		}
-		aggregate.workerCount++
-		aggregate.workerNames = append(aggregate.workerNames, worker.Name)
-		if locality := strings.TrimSpace(worker.ModelLocality); locality != "" {
-			aggregate.localities[locality] = struct{}{}
-			if aggregate.locality == "" {
-				aggregate.locality = locality
-			}
-			switch locality {
-			case interfaces.ModelLocalityLocal:
-				aggregate.localCount++
-			case interfaces.ModelLocalityCloud:
-				aggregate.cloudCount++
-			}
-		}
-
-		aggregate.capabilities = append(aggregate.capabilities, capabilityFromWorker(worker))
-		for _, operation := range worker.Operations {
-			mergeAggregateOperation(aggregate, operation)
-		}
-		collectAggregateResources(aggregate, worker, resourceByName, factoryCfg.Resources)
-	}
-
-	catalog := make(map[string]discoveredModelCatalogEntry, len(aggregates))
-	for key, aggregate := range aggregates {
-		sort.Strings(aggregate.workerNames)
-		sort.Slice(aggregate.capabilities, func(i, j int) bool {
-			return aggregate.capabilities[i].Worker < aggregate.capabilities[j].Worker
-		})
-		summary := buildModelSummary(*aggregate)
-		catalog[key] = discoveredModelCatalogEntry{
-			summary: summary,
-			detail: factoryapi.ModelDetail{
-				Name:             summary.Name,
-				ProviderLocality: summary.ProviderLocality,
-				Status:           summary.Status,
-				LoadState:        summary.LoadState,
-				Operations:       summary.Operations,
-				Modalities:       summary.Modalities,
-				Resources:        summary.Resources,
-				Capabilities:     aggregate.capabilities,
-				Diagnostics:      modelDiagnostics(*aggregate, summary),
-			},
-		}
-	}
-	return catalog
-}
-
-func capabilityFromWorker(worker interfaces.WorkerConfig) factoryapi.ModelCapability {
-	operations := make([]factoryapi.ModelOperation, 0, len(worker.Operations))
-	for _, operation := range worker.Operations {
-		operations = append(operations, generatedModelOperation(operation))
-	}
-	sort.Slice(operations, func(i, j int) bool {
-		return operations[i].Name < operations[j].Name
-	})
-	resourceNames := make([]string, 0, len(worker.Resources))
-	for _, resource := range worker.Resources {
-		if name := strings.TrimSpace(resource.Name); name != "" {
-			resourceNames = append(resourceNames, name)
-		}
-	}
-	sort.Strings(resourceNames)
-
-	capability := factoryapi.ModelCapability{
-		Worker:           worker.Name,
-		ProviderLocality: factoryapi.WorkerModelLocality(strings.TrimSpace(worker.ModelLocality)),
-		Operations:       operations,
-		ResourceNames:    resourceNames,
-	}
-	if provider := strings.TrimSpace(worker.ModelProvider); provider != "" {
-		generatedProvider := factoryapi.WorkerModelProvider(provider)
-		capability.ModelProvider = &generatedProvider
-	}
-	return capability
-}
-
-func mergeAggregateOperation(aggregate *discoveredModelAggregate, operation interfaces.ModelOperation) {
-	if aggregate == nil {
-		return
-	}
-	key := strings.TrimSpace(operation.Name)
-	if key == "" {
-		return
-	}
-	generated := generatedModelOperation(operation)
-	if existing, ok := aggregate.operations[key]; ok {
-		generated = mergeGeneratedOperations(existing, generated)
-	}
-	aggregate.operations[key] = generated
-	for _, slot := range operation.Inputs {
-		for _, contentType := range slot.ContentTypes {
-			if normalized := strings.TrimSpace(contentType); normalized != "" {
-				aggregate.modalities[factoryapi.ModelOperationContentType(normalized)] = struct{}{}
-			}
-		}
-	}
-	for _, slot := range operation.Outputs {
-		for _, contentType := range slot.ContentTypes {
-			if normalized := strings.TrimSpace(contentType); normalized != "" {
-				aggregate.modalities[factoryapi.ModelOperationContentType(normalized)] = struct{}{}
-			}
-		}
-	}
-}
-
-func mergeGeneratedOperations(left, right factoryapi.ModelOperation) factoryapi.ModelOperation {
-	merged := left
-	merged.Inputs = mergeGeneratedOperationSlots(left.Inputs, right.Inputs)
-	merged.Outputs = mergeGeneratedOperationSlots(left.Outputs, right.Outputs)
-	return merged
-}
-
-func mergeGeneratedOperationSlots(left, right *[]factoryapi.ModelOperationSlot) *[]factoryapi.ModelOperationSlot {
-	slotByName := make(map[string]factoryapi.ModelOperationSlot)
-	order := make([]string, 0)
-	appendSlots := func(slots *[]factoryapi.ModelOperationSlot) {
-		if slots == nil {
-			return
-		}
-		for _, slot := range *slots {
-			existing, ok := slotByName[slot.Name]
-			if !ok {
-				slotByName[slot.Name] = slot
-				order = append(order, slot.Name)
-				continue
-			}
-			slotByName[slot.Name] = mergeGeneratedOperationSlot(existing, slot)
-		}
-	}
-	appendSlots(left)
-	appendSlots(right)
-	if len(order) == 0 {
-		return nil
-	}
-	merged := make([]factoryapi.ModelOperationSlot, 0, len(order))
-	for _, name := range order {
-		merged = append(merged, slotByName[name])
-	}
-	return &merged
-}
-
-func mergeGeneratedOperationSlot(left, right factoryapi.ModelOperationSlot) factoryapi.ModelOperationSlot {
-	merged := left
-	typeSet := make(map[factoryapi.ModelOperationContentType]struct{}, len(left.ContentTypes)+len(right.ContentTypes))
-	ordered := make([]factoryapi.ModelOperationContentType, 0, len(left.ContentTypes)+len(right.ContentTypes))
-	for _, contentType := range left.ContentTypes {
-		if _, ok := typeSet[contentType]; ok {
-			continue
-		}
-		typeSet[contentType] = struct{}{}
-		ordered = append(ordered, contentType)
-	}
-	for _, contentType := range right.ContentTypes {
-		if _, ok := typeSet[contentType]; ok {
-			continue
-		}
-		typeSet[contentType] = struct{}{}
-		ordered = append(ordered, contentType)
-	}
-	merged.ContentTypes = ordered
-	if merged.Required == nil {
-		merged.Required = right.Required
-	} else if right.Required != nil && *right.Required {
-		required := true
-		merged.Required = &required
-	}
-	return merged
-}
-
-func collectAggregateResources(
-	aggregate *discoveredModelAggregate,
-	worker interfaces.WorkerConfig,
-	resourceByName map[string]interfaces.ResourceConfig,
-	allResources []interfaces.ResourceConfig,
-) {
-	if aggregate == nil {
-		return
-	}
-	for _, requirement := range worker.Resources {
-		resource, ok := resourceByName[requirement.Name]
-		if !ok {
-			continue
-		}
-		aggregate.hasAnyResource = true
-		if canonicalModelName(resource.Model) == canonicalModelName(worker.Model) && canonicalModelName(resource.Model) != "" {
-			aggregate.hasModelScoped = true
-		}
-		aggregate.resources[resource.Name] = generatedModelResourceSummary(resource)
-	}
-	for _, resource := range allResources {
-		if canonicalModelName(resource.Model) != canonicalModelName(worker.Model) {
-			continue
-		}
-		aggregate.hasAnyResource = true
-		aggregate.hasModelScoped = true
-		aggregate.resources[resource.Name] = generatedModelResourceSummary(resource)
-	}
-}
-
-func buildModelSummary(aggregate discoveredModelAggregate) factoryapi.ModelSummary {
-	operations := make([]factoryapi.ModelOperation, 0, len(aggregate.operations))
-	for _, operation := range aggregate.operations {
-		operations = append(operations, operation)
-	}
-	sort.Slice(operations, func(i, j int) bool {
-		return operations[i].Name < operations[j].Name
-	})
-
-	modalities := make([]factoryapi.ModelOperationContentType, 0, len(aggregate.modalities))
-	for modality := range aggregate.modalities {
-		modalities = append(modalities, modality)
-	}
-	sort.Slice(modalities, func(i, j int) bool {
-		return modalities[i] < modalities[j]
-	})
-
-	resources := make([]factoryapi.ModelResourceSummary, 0, len(aggregate.resources))
-	for _, resource := range aggregate.resources {
-		resources = append(resources, resource)
-	}
-	sort.Slice(resources, func(i, j int) bool {
-		return resources[i].Name < resources[j].Name
-	})
-
-	return factoryapi.ModelSummary{
-		Name:             aggregate.name,
-		ProviderLocality: factoryapi.WorkerModelLocality(primaryModelLocality(aggregate)),
-		Status:           modelStatus(aggregate),
-		LoadState:        modelLoadState(aggregate),
-		Operations:       operations,
-		Modalities:       modalities,
-		Resources:        resources,
-	}
-}
-
-func modelStatus(aggregate discoveredModelAggregate) factoryapi.ModelStatus {
-	if aggregate.localCount > 0 && !aggregate.hasModelScoped {
-		return factoryapi.ModelStatusUNAVAILABLE
-	}
-	return factoryapi.ModelStatusREADY
-}
-
-func modelLoadState(aggregate discoveredModelAggregate) factoryapi.ModelLoadState {
-	if primaryModelLocality(aggregate) == interfaces.ModelLocalityLocal {
-		return factoryapi.UNLOADED
-	}
-	return factoryapi.NOTAPPLICABLE
-}
-
-func primaryModelLocality(aggregate discoveredModelAggregate) string {
-	switch {
-	case aggregate.locality != "":
-		return aggregate.locality
-	case aggregate.localCount > 0:
-		return interfaces.ModelLocalityLocal
-	case aggregate.cloudCount > 0:
-		return interfaces.ModelLocalityCloud
-	default:
-		return interfaces.ModelLocalityCloud
-	}
-}
-
-func modelDiagnostics(aggregate discoveredModelAggregate, summary factoryapi.ModelSummary) factoryapi.StringMap {
-	diagnostics := factoryapi.StringMap{
-		"workerCount":      strconv.Itoa(aggregate.workerCount),
-		"localWorkerCount": strconv.Itoa(aggregate.localCount),
-		"cloudWorkerCount": strconv.Itoa(aggregate.cloudCount),
-		"resourceCount":    strconv.Itoa(len(summary.Resources)),
-		"workers":          strings.Join(aggregate.workerNames, ","),
-		"mixedLocality":    strconv.FormatBool(len(aggregate.localities) > 1),
-	}
-	if summary.Status == factoryapi.ModelStatusUNAVAILABLE {
-		diagnostics["statusReason"] = "local model workers require a matching MODEL resource declaration for readiness"
-	} else {
-		diagnostics["statusReason"] = "declared worker capabilities and resources are discoverable"
-	}
-	return diagnostics
-}
-
-func generatedModelOperation(operation interfaces.ModelOperation) factoryapi.ModelOperation {
-	generated := factoryapi.ModelOperation{
-		Name: strings.TrimSpace(operation.Name),
-	}
-	if len(operation.Inputs) > 0 {
-		inputs := make([]factoryapi.ModelOperationSlot, 0, len(operation.Inputs))
-		for _, slot := range operation.Inputs {
-			inputs = append(inputs, generatedModelOperationSlot(slot))
-		}
-		sort.Slice(inputs, func(i, j int) bool {
-			return inputs[i].Name < inputs[j].Name
-		})
-		generated.Inputs = &inputs
-	}
-	if len(operation.Outputs) > 0 {
-		outputs := make([]factoryapi.ModelOperationSlot, 0, len(operation.Outputs))
-		for _, slot := range operation.Outputs {
-			outputs = append(outputs, generatedModelOperationSlot(slot))
-		}
-		sort.Slice(outputs, func(i, j int) bool {
-			return outputs[i].Name < outputs[j].Name
-		})
-		generated.Outputs = &outputs
-	}
-	return generated
-}
-
-func generatedModelOperationSlot(slot interfaces.ModelOperationSlot) factoryapi.ModelOperationSlot {
-	contentTypes := make([]factoryapi.ModelOperationContentType, 0, len(slot.ContentTypes))
-	for _, contentType := range slot.ContentTypes {
-		if normalized := strings.TrimSpace(contentType); normalized != "" {
-			contentTypes = append(contentTypes, factoryapi.ModelOperationContentType(normalized))
-		}
-	}
-	sort.Slice(contentTypes, func(i, j int) bool {
-		return contentTypes[i] < contentTypes[j]
-	})
-	generated := factoryapi.ModelOperationSlot{
-		Name:         strings.TrimSpace(slot.Name),
-		ContentTypes: contentTypes,
-	}
-	if slot.Required {
-		required := true
-		generated.Required = &required
-	}
-	return generated
-}
-
-func generatedModelResourceSummary(resource interfaces.ResourceConfig) factoryapi.ModelResourceSummary {
-	summary := factoryapi.ModelResourceSummary{
-		Name:     resource.Name,
-		Type:     factoryapi.ResourceType(strings.TrimSpace(resource.Type)),
-		Capacity: resource.Capacity,
-	}
-	if value := strings.TrimSpace(resource.Model); value != "" {
-		summary.Model = &value
-	}
-	if value := strings.TrimSpace(resource.Backend); value != "" {
-		summary.Backend = &value
-	}
-	if value := strings.TrimSpace(resource.LoadPolicy); value != "" {
-		summary.LoadPolicy = &value
-	}
-	if value := strings.TrimSpace(resource.Provider); value != "" {
-		summary.Provider = &value
-	}
-	return summary
-}
-
-func canonicalModelName(model string) string {
-	return strings.ToUpper(strings.TrimSpace(model))
 }
 
 const directModelInvocationTransitionID = "direct-model-invocation"
@@ -688,7 +219,7 @@ func selectModelInvocationWorker(runtimeCfg *factoryconfig.LoadedFactoryConfig, 
 	if runtimeCfg == nil || runtimeCfg.FactoryConfig() == nil {
 		return nil, interfaces.ModelOperation{}, fmt.Errorf("runtime config is not available")
 	}
-	modelKey := canonicalModelName(modelName)
+	modelKey := localmodels.CanonicalModelName(modelName)
 	operationName = strings.TrimSpace(operationName)
 	if modelKey == "" {
 		return nil, interfaces.ModelOperation{}, fmt.Errorf("%w: empty model name", apisurface.ErrModelNotFound)
@@ -703,7 +234,7 @@ func selectModelInvocationWorker(runtimeCfg *factoryconfig.LoadedFactoryConfig, 
 		if !ok || workerDef == nil || workerDef.Type != interfaces.WorkerTypeModel {
 			continue
 		}
-		if canonicalModelName(workerDef.Model) != modelKey {
+		if localmodels.CanonicalModelName(workerDef.Model) != modelKey {
 			continue
 		}
 		modelMatched = true
