+++ pkg/service/testmain_test.go
@@ -14,7 +14,7 @@ import (
 	"github.com/portpowered/infinite-you/pkg/interfaces"
 	"github.com/portpowered/infinite-you/pkg/logging"
 	"github.com/portpowered/infinite-you/pkg/replay"
-	"github.com/portpowered/infinite-you/pkg/service/localmodel"
+	"github.com/portpowered/infinite-you/pkg/localmodels"
 	"github.com/portpowered/infinite-you/pkg/workers"
 	workerprovider "github.com/portpowered/infinite-you/pkg/workers/provider"
 	"go.uber.org/zap"
@@ -260,7 +260,7 @@ func localModelLongTestFactoryConfig() *interfaces.FactoryConfig {
 			Model:         "OMNIVOICE_Q4_K_M",
 			ModelProvider: interfaces.RunnerIDCodex,
 			ModelLocality: interfaces.ModelLocalityLocal,
-			Command:       localmodel.DefaultOmniVoiceCommand,
+			Command:       localmodels.DefaultOmniVoiceCommand,
 			Resources: []interfaces.ResourceConfig{{
 				Name:     "omnivoice-cache",
 				Capacity: 1,
